# Conflicts:
#	src/main/java/com/unifasservice/controller/ColorController.java
#	src/main/java/com/unifasservice/controller/ProductController.java
#	src/main/java/com/unifasservice/controller/SizeController.java
#	src/main/java/com/unifasservice/converter/ColorConverter.java
#	src/main/java/com/unifasservice/converter/SizeConverter.java
#	src/main/java/com/unifasservice/dto/payload/response/ColorResponse.java
#	src/main/java/com/unifasservice/dto/payload/response/SizeResponse.java
#	src/main/java/com/unifasservice/service/ColorService.java
#	src/main/java/com/unifasservice/service/ProductService.java
#	src/main/java/com/unifasservice/service/SizeService.java
#	src/main/java/com/unifasservice/service/UserService.java
#	src/main/java/com/unifasservice/service/impl/ColorServiceImpl.java
#	src/main/java/com/unifasservice/service/impl/ProductServiceImpl.java
#	src/main/java/com/unifasservice/service/impl/SizeServiceImpl.java
#	src/main/java/com/unifasservice/service/impl/UserServiceImpl.java

+++ src/main/java/com/unifasservice/dto/payload/response/OrderLineResponse.java
@@ -8,9 +8,9 @@ import lombok.Data;
 public class OrderLineResponse {
     private long id;
     private String productName;
-    private long price;
+    private double price;
     private int quantity;
-    private long subtotal;
+    private double subtotal;
     private String size;
     private String color;
 }
